Fit_Champs_Back/requirements.txt
alembic
bcrypt
cryptography
fastapi[standard]
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
psycopg2-binary